feat: finished task4 and project, reafctored any functions for project architecture.

+++ users/views.py
@@ -1,16 +1,12 @@
 from django.shortcuts import get_object_or_404
-
 from rest_framework.views import APIView, Request
 from rest_framework_simplejwt.authentication import JWTAuthentication
-
-
+from utils.methods import generate_response, generate_delete_response, generate_error_response
+from utils.permissions import EmployeeOrOwnerPermissions, AuthenticatedPermissions
+from utils.user import verify_user_exists
+from exceptions import UserException
 from .models import User
 from .serializers import UserSerializer
-from .utils import update_keys, verify_user_exists
-from .exceptions import UserException
-
-from utils.methods import generate_response, generate_delete_response, generate_error_response
-from utils.permissions import EmployeeOrOwnerPermissions
 
 
 class UserView(APIView):
@@ -37,28 +33,33 @@ class UserView(APIView):
 
 class UserDetailView(APIView):
     authentication_classes = [JWTAuthentication]
-    permission_classes = [EmployeeOrOwnerPermissions]
+    permission_classes = [AuthenticatedPermissions, EmployeeOrOwnerPermissions]
 
     def get(self, request: Request, user_id: int):
         user = get_object_or_404(User, id=user_id)
+
+        self.check_object_permissions(request, user)
+
         serializer = UserSerializer(user)
+
         return generate_response(200, serializer.data)
 
     def patch(self, request: Request, user_id: int):
         user = get_object_or_404(User, id=user_id)
-        serializer = UserSerializer(data=user)
+
+        self.check_object_permissions(request, user)
+
+        serializer = UserSerializer(user, request.data, partial=True)
         serializer.is_valid(raise_exception=True)
-
-        user_body = serializer.validated_data
-
-        update_keys(user_body.items(), user)
-
         serializer.save()
 
         return generate_response(200, serializer.data)
 
     def delete(self, request: Request, user_id: int):
         user = get_object_or_404(User, id=user_id)
+
+        self.check_object_permissions(request, user)
+
         user.delete()
 
         return generate_delete_response()
